More cleanup

+++ src/ExecutionContext.js
@@ -92,6 +92,8 @@ class ExecutionFrame {
         else if (frame.className) {
             if (typeof frame.className === 'string')
                 this.className = frame.className;
+            else if (frame.className.constructor)
+                this.className = frame.className.constructor.name;
             else
                 this.className = frame.className.name || 'Unknown';
         }
@@ -175,7 +177,7 @@ class ExecutionFrame {
      * How did this call originate?
      */
     get origin() {
-        let prev = this.context.length > 1 ? this.context.stack[1] : false,
+        let prev = this.context.length > 1 ? this.context._callstack[1] : false,
             val = CallOrigin.Unknown;
 
         if (prev) {
@@ -337,7 +339,7 @@ class ExecutionContext extends events.EventEmitter {
         this.controller = new ExecutionSignaller();
 
         /** @type {ExecutionFrame[]} */
-        this.stack = [];
+        this._callstack = [];
 
         /** @type {{ verb: string, args: string[] }[]} */
         this.cmdStack = [];
@@ -353,9 +355,9 @@ class ExecutionContext extends events.EventEmitter {
 
         if (parent) {
             /** @type {ExecutionFrame[]} */
-            this.stack = parent.stack.slice(0);
+            this._callstack = parent._callstack.slice(0);
             this.shell = parent.shell || false;
-            this.branchedAt = parent.stack.length;
+            this.branchedAt = parent._callstack.length;
             this.alarmTime = parent.alarmTime;
             this.currentVerb = parent.currentVerb || false;
             this.client = parent.client;
@@ -363,7 +365,7 @@ class ExecutionContext extends events.EventEmitter {
 
             if (createDetached !== true) {
                 this.parent = parent;
-                this.parentFrame = parent.stack[0];
+                this.parentFrame = parent._callstack[0];
 
                 parent.addNewChild(this);
 
@@ -571,7 +573,7 @@ class ExecutionContext extends events.EventEmitter {
      */
     async awaitResult(asyncCode) {
         let startTime = new Date().getTime(),
-            frame = this.stack[0];
+            frame = this._callstack[0];
 
         if (!frame)
             await driver.crashAsync(new Error(`Call stack was empty!`));
@@ -622,9 +624,9 @@ class ExecutionContext extends events.EventEmitter {
 
         let result = new ExecutionContext(this, false, info);
         if (lineNumber > 0) {
-            let lastFrame = result.stack[0].clone();
+            let lastFrame = result._callstack[0].clone();
             lastFrame.lineNumber = lineNumber;
-            result.stack[0] = lastFrame;
+            result._callstack[0] = lastFrame;
         }
         return result;
     }
@@ -668,13 +670,15 @@ class ExecutionContext extends events.EventEmitter {
      */
     complete(forceCompletion = false) {
         try {
-            if (this.stack.length === this.branchedAt || true === forceCompletion) {
+            if (this._callstack.length === this.branchedAt || true === forceCompletion) {
                 if (this.childCount > 0) {
                     for (const [id, child] of Object.entries(this.children)) {
-                        console.log(`\tExecutionContext: Child context ${id} has not completed`);
-                        child.onComplete.push(() => {
-                            console.log(`\tExecutionContext: Child context ${id} completed after its parent`);
-                        });
+                        if (global.driver?.debugMode) {
+                            console.log(`\tExecutionContext: Child context ${id} has not completed`);
+                            child.onComplete.push(() => {
+                                console.log(`\tExecutionContext: Child context ${id} completed after its parent`);
+                            });
+                        }
                         if (child.handleId in contexts) {
                             delete contexts[child.handleId];
                         }
@@ -682,8 +686,10 @@ class ExecutionContext extends events.EventEmitter {
                 }
                 if (this.unusedChildCount > 0) {
                     for (const [id, child] of Object.entries(this.unusedChildren)) {
-                        console.log(`\tExecutionContext: Child context ${child.handleId} was never used`);
-                        ExecutionContext.deleteContext(child);
+                        if (global.driver?.debugMode) {
+                            console.log(`\tExecutionContext: Child context ${child.handleId} was never used`);
+                            ExecutionContext.deleteContext(child);
+                        }
                     }
                 }
                 this.completed = true;
@@ -730,7 +736,7 @@ class ExecutionContext extends events.EventEmitter {
     }
 
     get currentFileName() {
-        let frame = this.stack[0];
+        let frame = this._callstack[0];
         return frame.file || false;
     }
 
@@ -801,15 +807,15 @@ class ExecutionContext extends events.EventEmitter {
      * @param {number} index The index of the frame to fetch
      */
     getFrame(index) {
-        return index > -1 && index < this.length && this.stack[index];
+        return index > -1 && index < this.length && this._callstack[index];
     }
 
     /**
      * Returns the last frame that was internal to the game.
      */
     getLastGameFrame() {
-        let lastFrame = this.stack.lastIndexOf(frame => ExecutionContext.isExternalPath(frame.file));
-        return lastFrame > -1 && this.stack[lastFrame]
+        let lastFrame = this._callstack.lastIndexOf(frame => ExecutionContext.isExternalPath(frame.file));
+        return lastFrame > -1 && this._callstack[lastFrame]
     }
 
     /**
@@ -820,7 +826,7 @@ class ExecutionContext extends events.EventEmitter {
      * @returns {string}
      */
     getStackString(startFrame = 0, stopFrame = undefined, maskExternalPaths = true) {
-        return this.stack.slice(startFrame, stopFrame).map(f => f.toString(maskExternalPaths)).join('\n');
+        return this._callstack.slice(startFrame, stopFrame).map(f => f.toString(maskExternalPaths)).join('\n');
     }
 
     /**
@@ -878,7 +884,7 @@ class ExecutionContext extends events.EventEmitter {
     }
 
     get isAwaited() {
-        return this.stack.length > 0 && this.stack[0].awaitCount > 0;
+        return this._callstack.length > 0 && this._callstack[0].awaitCount > 0;
     }
 
     /**
@@ -897,7 +903,7 @@ class ExecutionContext extends events.EventEmitter {
     }
 
     get length() {
-        return this.stack.length;
+        return this._callstack.length;
     }
 
     /**
@@ -934,12 +940,11 @@ class ExecutionContext extends events.EventEmitter {
      * @param {boolean} yieldBack If true, then the call duration is yielded back to the alarm time
      */
     pop(method, yieldBack = false) {
-        let lastFrame = this.stack.shift();
+        let lastFrame = this._callstack.shift();
 
         if (!lastFrame || lastFrame.callString !== method) {
-            if (lastFrame && lastFrame !== method) {
+            if (lastFrame && lastFrame !== method)
                 console.log(`\tExecutionContext: Out of sync; Expected ${method} but found ${lastFrame.callString}`);
-            }
             else
                 console.log(`\tExecutionContext: Out of sync... no frames left!`);
         }
@@ -949,7 +954,7 @@ class ExecutionContext extends events.EventEmitter {
                 this.alarmTime += lastFrame.ellapsed;
         }
 
-        if (this.stack.length === this.branchedAt) {
+        if (this._callstack.length === this.branchedAt) {
             this.complete();
         }
         else
@@ -963,7 +968,7 @@ class ExecutionContext extends events.EventEmitter {
     }
 
     popCurrentFrame() {
-        let frame = this.stack[0];
+        let frame = this._callstack[0];
         return this.pop(frame.callString);
     }
 
@@ -994,27 +999,25 @@ class ExecutionContext extends events.EventEmitter {
             if (this.alarmTime < Number.MAX_SAFE_INTEGER)
                 this.alarmTime += lastFrame.ellapsed;
         }
-        if (frame !== this.stack[0]) {
-            if (this.stack.length === 0) {
-                //  Crasher?
-                console.log('\tExecutionContext: Cannot pop frame off an empty stack!');
+        if (frame !== this._callstack[0]) {
+            if (this._callstack.length === 0) {
+                global.driver?.crash(new Error(`ExecutionContext: Cannot pop frame off an empty stack! (frame: ${frame})`));
             }
             else {
-                console.log(`\tExecutionContext: Out of sync; Expected ${frame.method} but found ${this.stack[0].method}`);
-                let index = this.stack.findIndex(f => f === frame);
+                console.log(`\tExecutionContext: Out of sync; Expected ${frame.method} but found ${this._callstack[0].method}`);
+                let index = this._callstack.findIndex(f => f === frame);
                 if (index === -1) {
-                    //  Crash?
-                    console.log(`\tExecutionContext: CRITICAL ERROR: FRAME NOT FOUND IN STACK: ${frame.method}`);
+                    global.driver?.crash(new Error(`ExecutionContext: CRITICAL ERROR: FRAME NOT FOUND IN STACK! (frame: ${frame})`));
                 }
                 else {
-                    this.stack = this.stack.slice(0, index);
+                    this._callstack = this._callstack.slice(0, index);
                 }
             }
         }
         else
-            this.stack.shift();
+            this._callstack.shift();
 
-        if (this.stack.length <= this.branchedAt) {
+        if (this._callstack.length <= this.branchedAt) {
             this.complete();
         }
         else
@@ -1052,7 +1055,7 @@ class ExecutionContext extends events.EventEmitter {
      * @type {typeof import('./MUDObject')[]}
      */
     get previousObjects() {
-        let stack = this.stack;
+        let stack = this._callstack;
         return stack
             .filter(f => typeof f.object === 'object')
             .slice(0)
@@ -1069,7 +1072,7 @@ class ExecutionContext extends events.EventEmitter {
         else if (typeof frameInfo.method !== 'string')
             driver.crash(new Error('CRASH: pushFrameObject() received invalid parameter'));
         let newFrame = new ExecutionFrame({ context: this, ...frameInfo });
-        this.stack.unshift(newFrame);
+        this._callstack.unshift(newFrame);
         if (!this.used) {
             this.used = true;
             this.parent?.addActiveChild(this);
@@ -1104,9 +1107,9 @@ class ExecutionContext extends events.EventEmitter {
      * @param {string} frameId The frame's UUID
      */
     removeFrameById(frameId) {
-        let index = this.stack.findIndex(frame => frame.id === frameId);
+        let index = this._callstack.findIndex(frame => frame.id === frameId);
         if (index > -1) {
-            let result = this.stack.splice(index, 1);
+            let result = this._callstack.splice(index, 1);
             return result.length > 0;
         }
         return false;
@@ -1161,8 +1164,8 @@ class ExecutionContext extends events.EventEmitter {
      * @type {typeof import('./MUDObject')}
      */
     get thisObject() {
-        for (let i = 0, m = this.stack.length; i < m; i++) {
-            let ob = this.stack[i].object;
+        for (let i = 0, m = this._callstack.length; i < m; i++) {
+            let ob = this._callstack[i].object;
 
             if (typeof ob === 'object')
                 return ob;
@@ -1297,7 +1300,7 @@ class ExecutionContext extends events.EventEmitter {
 
     validSyncCall(filename, lineNumber, expr) {
         let ecc = ExecutionContext.getCurrentExecution(),
-            frame = ecc && ecc.stack[0] || false;
+            frame = ecc && ecc._callstack[0] || false;
 
         let result = expr();
         if (driver.efuns.isPromise(this, result)) {
